No functional changes, refactor only
