--- test/sample_feature_list.yml
@@ -1,8 +1,0 @@
-development:
-  my_awesome_feature: enabled 
-
-stage:
-  my_awesome_feature: whitelist_only
-
-prod:
-  my_awesome_feature: disabled
